% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idfEEZ.R
\docType{data}
\name{idfEEZ}
\alias{idfEEZ}
\title{Recognize EEZ region Function
This function allows you to identify location in which EEZ from a numeric vector.}
\format{
list with random numbers
}
\usage{
idfEEZ

idfEEZ(data, lon, lat)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Recognize EEZ region Function
This function allows you to identify location in which EEZ from a numeric vector.
}
\examples{
idfEEZ()
}
\keyword{location}
